+++ cmn-data-svc-server/src/main/java/org/acumos/cds/repository/SolTagMapRepository.java
@@ -22,7 +22,6 @@ package org.acumos.cds.repository;
 
 import org.acumos.cds.domain.MLPSolTagMap;
 import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
@@ -37,7 +36,6 @@ public interface SolTagMapRepository extends CrudRepository<MLPSolTagMap, MLPSol
 	 */
 	@Modifying
 	@Transactional // throws exception without this
-	@Query(value = "DELETE FROM MLPSolTagMap m WHERE m.solutionId = :solutionId")
-	void deleteTagsForSolution(@Param("solutionId") String solutionId);
+	void deleteBySolutionId(@Param("solutionId") String solutionId);
 
 }